Added message pro processor
