+++ DFC.App.JobProfileTasks.MessageFunctionApp/Services/IMessagePreProcessor.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Azure.ServiceBus;
+using System.Threading.Tasks;
+
+namespace DFC.App.JobProfileTasks.MessageFunctionApp.Services
+{
+    public interface IMessagePreProcessor
+    {
+        Task Process(Message sitefinityMessage);
+    }
+}
